+++ react/__mocks__/geolocationAddressValid.js
@@ -1,0 +1,16 @@
+export default {
+  addressId: {value: '3095115299110'},
+  addressQuery: {value: 'Praia de Botafogo - Botafogo, Rio de Janeiro - RJ, 22250-040, Brasil', geolocationAutoCompleted: true},
+  addressType: {value: 'residential'},
+  city: {value: 'Rio de Janeiro',  notApplicable: false, geolocationAutoCompleted: true},
+  complement: {value: null},
+  country: {value: 'BRA', geolocationAutoCompleted: true},
+  geoCoordinates: {visited: true, value: [], geolocationAutoCompleted: true},
+  neighborhood: {value: 'Botafogo',  notApplicable: false, geolocationAutoCompleted: true},
+  number: {focus: true,  notApplicable: true, valid: true, loading: false, reason: null},
+  postalCode: {value: '22250-040', postalCodeAutoCompleted: true, loading: false},
+  receiverName: {value: 'auto auto'},
+  reference: {value: null},
+  state: {value: 'RJ',  notApplicable: false, geolocationAutoCompleted: true},
+  street: {value: 'Praia de Botafogo',  notApplicable: false, geolocationAutoCompleted: true},
+}