+++ react/country/CHL.js
@@ -502,7 +502,7 @@ export default {
       types: ['postal_code'],
       required: false,
     },
-    number: { valueIn: 'long_name', types: ['street_number'], required: false },
+    number: { valueIn: 'long_name', types: ['street_number'], required: false, notApplicable: true },
     street: { valueIn: 'long_name', types: ['route'], required: false },
     neighborhood: {
       valueIn: 'long_name',