+++ react/country/COL.js
@@ -1305,7 +1305,12 @@ export default {
         return address
       },
     },
-    number: { valueIn: 'long_name', types: ['street_number'], required: false },
+    number: {
+      valueIn: 'long_name',
+      types: ['street_number'],
+      required: false,
+      notApplicable: true,
+    },
     street: {
       valueIn: 'long_name',
       types: ['route'],