+++ react/country/ROU.js
@@ -13987,7 +13987,7 @@ export default {
       types: ['postal_code'],
       required: false,
     },
-    number: { valueIn: 'long_name', types: ['street_number'], required: false },
+    number: { valueIn: 'long_name', types: ['street_number'], required: false, notApplicable: true },
     street: { valueIn: 'long_name', types: ['route'], required: true },
     state: {
       valueIn: 'long_name',