+++ react/country/__mocks__/usePostalCode.js
@@ -82,6 +82,11 @@ export default {
       valueIn: 'long_name',
       types: ['street_number'],
       required: false,
+      notApplicable: true,
+      handler: address => {
+        return {...address,
+          number: {...address['number'], notApplicable: true}}
+      },
     },
     street: {
       valueIn: 'long_name',