+++ react/helpers.js
@@ -2,6 +2,7 @@ import { addValidation, removeValidation } from './transforms/address'
 import { injectRules } from './addressRulesContext'
 import { injectAddressContext } from './addressContainerContext'
 import { isValidAddress, validateField, validateAddress } from './validateAddress'
+import getAddressByGeolocation from './geolocation/Utils'
 
 export default {
   isValidAddress,
@@ -11,4 +12,5 @@ export default {
   injectRules,
   addValidation,
   removeValidation,
+  getAddressByGeolocation,
 }