+++ react/propTypes/AddressShapeWithValidation.js
@@ -16,8 +16,10 @@ export const AddressWithValidation = reduce(
       ...ValidationType,
       autoCompleted: PropTypes.bool,
       loading: PropTypes.bool,
+      disabled: PropTypes.bool,
       value: valueType,
       valueOptions: PropTypes.arrayOf(PropTypes.string),
+      notApplicable: PropTypes.bool,
     }).isRequired
     return acc
   },